*: prepare v3 branch
github.com/simonpasquier/klog-gokit/v3 will only support k8s.io/klog/v2
with github.com/go-kit/kit.
