+++ README.md
@@ -1,39 +1,32 @@
 # klog-gokit [![CircleCI](https://circleci.com/gh/simonpasquier/klog-gokit.svg?style=svg)](https://circleci.com/gh/simonpasquier/klog-gokit)
 
-This package is a replacement for [k8s.io/klog](https://github.com/kubernetes/klog)
-in projects that use the [go-kit logger](https://godoc.org/github.com/go-kit/log).
+This package is a replacement for [k8s.io/klog/v2](https://github.com/kubernetes/klog)
+in projects that use the [github.com/go-kit/log](https://pkg.go.dev/github.com/go-kit/log) module for logging.
 
-It also supports [k8s.io/klog/v2](https://pkg.go.dev/k8s.io/klog/v2).
+*The current branch supports neither
+[k8s.io/klog](https://pkg.go.dev/k8s.io/klog) nor
+[github.com/go-kit/kit](https://pkg.go.dev/github.com/go-kit/kit). Please use the `v2.1.0`
+version instead.*
 
 It is heavily inspired by the [`github.com/kubermatic/glog-gokit`](https://github.com/kubermatic/glog-gokit) package.
 
 ## Usage
 
-Override the official klog package with this one.
-This simply replaces the code in `vendor/k8s.io/klog` with the code of this package.
-
-**With dep**
-
-In your `Gopkg.toml`:
-```toml
-[[override]]
-  name = "k8s.io/klog"
-  source = "github.com/simonpasquier/klog-gokit"
-```
-
-**With Go modules**
-
 Add this line to your `go.mod` file:
 
 ```
-replace k8s.io/klog => github.com/simonpasquier/klog-gokit master
+replace k8s.io/klog/v2 => github.com/simonpasquier/klog-gokit/v3 v3
 ```
 
 In your `main.go`:
 ```go
 // Import the package like it is original klog
-import "k8s.io/klog"
-
+import (
+    ...
+    "github.com/go-kit/log"
+    klog "k8s.io/klog/v2"
+    ...
+)
 
 // Create go-kit logger in your main.go
 logger := log.NewLogfmtLogger(log.NewSyncWriter(os.Stdout))
@@ -46,7 +39,8 @@ logger = level.NewFilter(logger, level.AllowAll())
 klog.SetLogger(logger)
 ```
 
-Setting the logger to the klog package **MUST** happen before using klog in any package.
+Setting the klog's logger **MUST** happen at the very beginning of your program
+(e.g. before using the other klog functions).
 
 ## Function Levels
 
@@ -56,6 +50,8 @@ Setting the logger to the klog package **MUST** happen before using klog in any 
 | InfoDepth    | Debug |
 | Infof        | Debug |
 | Infoln       | Debug |
+| InfoS        | Debug |
+| InfoSDepth   | Debug |
 | Warning      | Warn  |
 | WarningDepth | Warn  |
 | Warningf     | Warn  |
