+++ go.mod
@@ -1,11 +1,8 @@
-module github.com/simonpasquier/klog-gokit
+module github.com/simonpasquier/klog-gokit/v3
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-kit/log v0.1.0
-	github.com/kr/pretty v0.1.0 // indirect
 	github.com/stretchr/testify v1.7.0
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 )
 
 go 1.12
