--- v2/go.mod
@@ -1,5 +1,0 @@
-module github.com/simonpasquier/klog-gokit/v2
-
-require github.com/go-kit/log v0.1.0
-
-go 1.12
